Process ack batch migration complete messages in correct order
We process a message received from another worker confirming
that it has sent all steps for this migration phase by sending
it through the same code path as the immigrant step messages.
This way, we know that we've processed the last of those migration
message before we confirm that we are ready to receive data.